requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.65.0
scipy>=1.10.0
shapely>=2.0.0
pyclipper>=1.3.0

# Optional: Spell checking
pyspellchecker>=0.7.0

